+++ src/catkin_projects/drake_iiwa_sim/src/joint_space_trajectory_plan.cc
@@ -1,0 +1,1 @@
+#include <robot_plan_runner/joint_space_trajectory_plan.h>
