--- src/catkin_projects/drake_iiwa_sim/src/plan_types.cc
@@ -1,2 +1,0 @@
-#include "plan_types.h"
-
